[moz] Remove ETW from the build.
Fixing the ETW dependency is probably doable but it doesn't seem straightforward
with our cross-compilation setup and having a few ETW markers in DXC isn't
hugely important for us so this commit removes them from the build.

To reduce risk of rebase conflicts, the DxcEtw API is stubbed out in
include/dxcetw.h which would have normally been generated as part of the build.