[moz] Workaround to disable SPIRV codegen.
For some reason this CMakeLists file is not picking up on the fact that we are cross-compiling.
This causes it to assume that we are targetting linux instead of windows and unconditionally
include SPIRV in the build, which we do not want.
As a workaround until we figure out how to get the build to understand that we are targetting
Windows, comment out the part that enforces SPIRV to be in the build.